Improves code readability, CLI parameterization, user feedback and fixes several bugs
The CLI parameters have been simplified to provide a more clear interface for the users,
the output is now based around the output directories or files being set, if none of
those are present, no files will be processed. A logging instance has been implemented
that can supress or output more information, the SlideStandardization debug images
are now written depending on the output level handled by the logging instance. A small
bug that interpreted non-TIF files as TIF has been dealt with, and the user receives
feedback while writing the WSI to the disk, as to avoid any assumptions about the
program being frozen. A large amount of methods have had their parameterization
improved where each reference and parameter is more explicit about their usage of the
passed variable.

- Improves and cleans CLI parameterization
- Improves code readability by providing better method parameterization.
- Fixes any bugs in regards to TIF vs non-TIF files.